refactor: renaming input io module using pep8 convention.
